Generalize isolated FREEZER test fixtures
